requirements.txt
flask
line-bot-sdk>=3.0
Pillow
gunicorn
google-generativeai
